Reorganize directories for refractor
